+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [feat]: `round` method;
 - [feat]: `floor` method;
 - [feat]: `ceil` method;
+- [feat]: `clamp` method;
 
 ## 0.13.0 (2022-12-10)